+++ src/main/java/com/finalpre/quickshare/service/StoragePolicyService.java
@@ -1,0 +1,5 @@
+package com.finalpre.quickshare.service;
+
+public interface StoragePolicyService {
+    StoragePolicy getPolicy();
+}